#172 Add show slideshow button
