+++ src/main/resources/appresource.properties
@@ -273,7 +273,8 @@ knowledge.edit.label.delete.upload=Remove file
 knowledge.edit.label.fail.delete.upload=Fail remove file
 knowledge.edit.label.confirm.delete=Are you sure you want to delete?
 knowledge.edit.noaccess=You can't edit this knowledge
-knowledge.edit.set.image.path=Set the display path of image
+knowledge.edit.set.image.path=Add Image in this contents
+knowledge.edit.set.slide.path=Add SlideShow in this contents
 knowledge.delete.none=Knowledge to be deleted is not specified
 
 knowledge.likes.title=User list that was me press the "Like"
