+++ src/main/resources/appresource_ja.properties
@@ -273,7 +273,8 @@ knowledge.edit.label.delete.upload=ファイルを削除しました
 knowledge.edit.label.fail.delete.upload=ファイルの削除に失敗しました
 knowledge.edit.label.confirm.delete=本当に削除しますか？
 knowledge.edit.noaccess=編集権限がありません
-knowledge.edit.set.image.path=添付した画像を表示する
+knowledge.edit.set.image.path=画像を表示
+knowledge.edit.set.slide.path=スライドショー表示
 knowledge.delete.none=削除するナレッジが指定されていません
 
 knowledge.likes.title=「いいね！」を押してくれたユーザ一覧
