+++ src/main/webapp/js/common.js
@@ -15,3 +15,18 @@ $(document).ready(function() {
     });
 
 });
+function insertAtCaret(target, str) {
+    var obj = $(target);
+    obj.focus();
+    if (navigator.userAgent.match(/MSIE/)) {
+        var r = document.selection.createRange();
+        r.text = str;
+        r.select();
+    } else {
+        var s = obj.val();
+        var p = obj.get(0).selectionStart;
+        var np = p + str.length;
+        obj.val(s.substr(0, p) + str + s.substr(p));
+        obj.get(0).setSelectionRange(np, np);
+    }
+}
